meditrack-backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
supabase>=2.3.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
pytesseract>=0.3.10
Pillow>=10.2.0
openai>=1.10.0
twilio>=8.11.1
firebase-admin>=6.4.0
boto3>=1.34.34
python-multipart>=0.0.6
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
email-validator>=2.1.0
